# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "个人博客后端 API 服务，所有业务接口均使用统一响应结构。",
        "title": "Blog Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/admin/categories": {
            "post": {
                "description": "管理员新增分类；父分类字段仅预留，当前不处理父子层级。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "新增分类",
                "parameters": [
                    {
                        "description": "分类新增请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示新增成功，code=400 表示参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryItemResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}": {
            "put": {
                "description": "管理员修改分类名称、排序、图标、描述、封面等属性。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "修改分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "分类修改请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示修改成功，code=400 表示参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryItemResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "分类不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员删除分类，并清空该分类下文章关联；文章本身不会被删除。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "删除分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示删除成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryItemResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "分类不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/admin/categories/{id}/articles": {
            "post": {
                "description": "管理员把单篇或多篇文章加入分类；文章原有分类关系会迁移到当前分类。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "添加文章到分类",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文章 ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryArticlesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示添加成功，code=400 表示参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryArticlesResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "分类或文章不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "delete": {
                "description": "管理员批量移除分类下文章关联；文章本身不会被删除。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "删除分类下文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "分类 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "文章 ID 列表",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CategoryArticlesReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示移除成功，code=400 表示参数错误",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryArticlesResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "未登录或 token 已过期",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "403": {
                        "description": "权限不足",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "404": {
                        "description": "分类不存在",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/categories": {
            "get": {
                "description": "返回所有分类及其公开文章数量，按 seq ASC、文章数量 DESC 排序。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "分类"
                ],
                "summary": "查询分类 Tab 列表",
                "responses": {
                    "200": {
                        "description": "统一响应；code=0 表示查询成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryTabsResp"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategoryArticlesReq": {
            "type": "object",
            "required": [
                "article_ids"
            ],
            "properties": {
                "article_ids": {
                    "description": "ArticleIDs 文章 ID 列表；单篇文章也使用一个元素的数组。",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        1
                    ]
                }
            }
        },
        "dto.CategoryArticlesResp": {
            "type": "object",
            "properties": {
                "affected_count": {
                    "description": "AffectedCount 实际影响的关联数量。",
                    "type": "integer",
                    "example": 2
                },
                "article_ids": {
                    "description": "ArticleIDs 本次请求归一化后的文章 ID 列表。",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "category_id": {
                    "description": "CategoryID 分类 ID。",
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "dto.CategoryCreateReq": {
            "type": "object",
            "required": [
                "cover_img_url",
                "description",
                "icon",
                "name",
                "seq"
            ],
            "properties": {
                "cover_img_url": {
                    "description": "CoverImgUrl 封面图地址或对象 key。",
                    "type": "string",
                    "example": "covers/programming.jpg"
                },
                "description": {
                    "description": "Description 分类描述。",
                    "type": "string",
                    "example": "编程学习与工程实践"
                },
                "icon": {
                    "description": "Icon 图标地址或对象 key。",
                    "type": "string",
                    "example": "icons/programming.svg"
                },
                "name": {
                    "description": "Name 分类名称。",
                    "type": "string",
                    "example": "编程"
                },
                "parent_id": {
                    "description": "ParentID 父分类 ID，当前仅预留，不参与层级逻辑。",
                    "type": "integer",
                    "example": 1
                },
                "seq": {
                    "description": "Seq 排序值，越小越靠前；0 是有效值，因此用指针区分未传。",
                    "type": "integer",
                    "example": 0
                },
                "url": {
                    "description": "URL 分类路由别名。",
                    "type": "string",
                    "example": "programming"
                }
            }
        },
        "dto.CategoryItemResp": {
            "type": "object",
            "properties": {
                "article_count": {
                    "description": "ArticleCount 该分类下的公开文章数量。",
                    "type": "integer",
                    "example": 12
                },
                "cover_img_url": {
                    "description": "CoverImgUrl 封面图地址或对象 key。",
                    "type": "string"
                },
                "description": {
                    "description": "Description 分类描述。",
                    "type": "string"
                },
                "icon": {
                    "description": "Icon 图标 URL 或对象 key。",
                    "type": "string"
                },
                "id": {
                    "description": "ID 分类 ID。",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name 分类名称。",
                    "type": "string",
                    "example": "编程"
                },
                "parent_id": {
                    "description": "ParentID 父分类 ID，当前仅预留。",
                    "type": "integer",
                    "example": 1
                },
                "seq": {
                    "description": "Seq 排序值，越小越靠前。",
                    "type": "integer",
                    "example": 0
                },
                "url": {
                    "description": "URL 分类路由别名。",
                    "type": "string",
                    "example": "programming"
                }
            }
        },
        "dto.CategoryTabItem": {
            "type": "object",
            "properties": {
                "article_count": {
                    "description": "ArticleCount 该分类下的公开文章数量。",
                    "type": "integer",
                    "example": 12
                },
                "cover_img_url": {
                    "description": "CoverImgUrl 封面图地址。",
                    "type": "string"
                },
                "description": {
                    "description": "Description 分类描述。",
                    "type": "string"
                },
                "icon": {
                    "description": "Icon 图标 URL。",
                    "type": "string"
                },
                "id": {
                    "description": "ID 分类 ID。",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Name 分类名称。",
                    "type": "string",
                    "example": "编程"
                },
                "seq": {
                    "description": "Seq 排序值，越小越靠前。",
                    "type": "integer",
                    "example": 0
                },
                "url": {
                    "description": "URL 分类路由别名。",
                    "type": "string",
                    "example": "programming"
                }
            }
        },
        "dto.CategoryTabsResp": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "List 分类列表，按 seq ASC、文章数量 DESC 排序。",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryTabItem"
                    }
                }
            }
        },
        "dto.CategoryUpdateReq": {
            "type": "object",
            "properties": {
                "cover_img_url": {
                    "description": "CoverImgUrl 封面图地址或对象 key；传空字符串表示清空。",
                    "type": "string",
                    "example": "covers/programming.jpg"
                },
                "description": {
                    "description": "Description 分类描述；传空字符串表示清空。",
                    "type": "string",
                    "example": "编程学习与工程实践"
                },
                "icon": {
                    "description": "Icon 图标地址或对象 key；传空字符串表示清空。",
                    "type": "string",
                    "example": "icons/programming.svg"
                },
                "name": {
                    "description": "Name 分类名称。",
                    "type": "string",
                    "example": "编程"
                },
                "parent_id": {
                    "description": "ParentID 父分类 ID，当前仅预留，不参与层级逻辑。",
                    "type": "integer",
                    "example": 1
                },
                "seq": {
                    "description": "Seq 排序值，越小越靠前。",
                    "type": "integer",
                    "example": 0
                },
                "url": {
                    "description": "URL 分类路由别名；传空字符串表示清空。",
                    "type": "string",
                    "example": "programming"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "0 表示成功，非 0 为业务错误码",
                    "type": "integer"
                },
                "data": {
                    "description": "失败时省略，不输出 null"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}